refactor(task-page): split into smaller components
Extract task page into focused, reusable components and utilities:
- TaskPageHeader: PR/branch info, editor dropdown, title
- TaskPageMessageList: scrollable timeline with messages and activity
- TaskPageInput: textarea, model picker, send button
- CollapsedActivityGroup: collapsible tool call groups
- task-timeline.ts: timeline types and building logic
- task-activity-mapper.ts: event-to-activity mapping and detail formatting
- fail-task-run.ts: task run failure API call
- format-duration.ts: duration formatting
- pull-request.ts: PR display helpers (getters/formatters)

Main task-page.tsx now focuses solely on orchestration: state management,
effects, event handling, and component composition. Reduces file from 1,477
to 331 lines.
